Setup common-scripts package
