+++ package.json
@@ -31,7 +31,7 @@
   },
   "devDependencies": {
     "lerna": "^3.15.0",
-    "prettier": "^1.17.1",
-    "typescript": "^3.5.1"
+    "prettier": "^1.18.2",
+    "typescript": "^3.5.3"
   }
 }
